update docs line numbers

+++ src/neighbor.rs
@@ -8,7 +8,7 @@ use std::collections::HashMap;
 use crate::particle::Particle;
 
 /// 三维网格单元索引
-type Cell = (i32, i32, i32);
+type Cell = (i64, i64, i64);
 
 /// 为所有粒子构建邻域列表。
 ///
@@ -31,7 +31,11 @@ pub fn build_neighbor_list(particles: &[Particle], radius: f64) -> Vec<Vec<usize
         for dx in -1..=1 {
             for dy in -1..=1 {
                 for dz in -1..=1 {
-                    let cell = (base.0 + dx, base.1 + dy, base.2 + dz);
+                    let cell = (
+                        base.0 + dx as i64,
+                        base.1 + dy as i64,
+                        base.2 + dz as i64,
+                    );
                     if let Some(list) = grid.get(&cell) {
                         for &j in list {
                             if i == j {
@@ -54,9 +58,9 @@ pub fn build_neighbor_list(particles: &[Particle], radius: f64) -> Vec<Vec<usize
 
 fn cell_index(pos: [f64; 3], size: f64) -> Cell {
     (
-        (pos[0] / size).floor() as i32,
-        (pos[1] / size).floor() as i32,
-        (pos[2] / size).floor() as i32,
+        (pos[0] / size).floor() as i64,
+        (pos[1] / size).floor() as i64,
+        (pos[2] / size).floor() as i64,
     )
 }
